refactor: split video and code scripts
