+++ .github/workflows/02-tests.yml
@@ -20,7 +20,7 @@ jobs:
           uv pip install --system -r requirements.txt
       - name: Run tests
         run: |
-          pytest --cov=./scripts --cov=./tests
+          pytest --cov=./src --cov=./tests
       - name: Upload coverage reports to Codecov
         uses: codecov/codecov-action@v5
         with:
