+++ .github/workflows/contrib-heatmap.yml
@@ -23,7 +23,7 @@ jobs:
       - name: Generate heat-map
         env:
           GH_TOKEN: ${{ secrets.GH_TOKEN }}
-        run: python scripts/generate_heatmap.py
+        run: python src/generate_heatmap.py
       - name: Commit and push
         uses: EndBug/add-and-commit@v9
         with:
