+++ AGENTS.md
@@ -12,20 +12,21 @@ Avoid adding setup or asset instructions there; link to INSTRUCTIONS instead.
 
 | Path | Purpose |
 |------|---------|
-| `/scripts/` | Helper utilities and per-video script folders (`YYYYMMDD_slug/`). Each folder may include `script.md`, `metadata.json`, and optional `sources.txt` or `footage.md`. Run `scaffold_videos.py` to fetch titles/dates and create folders from `video_ids.txt`. |
+| `/src/` | Helper utilities and CLI tools. |
+| `/video_scripts/` | Per-video folders (`YYYYMMDD_slug/`) containing `script.md`, `metadata.json`, and optional `sources.txt` or `footage.md`. Run `scaffold_videos.py` to fetch titles/dates and create folders from `video_ids.txt`. |
 | `/ideas/` | Checklist-style idea files – raw, WIP content. |
 | `/schemas/` | JSON-Schemas (e.g. `video_metadata.schema.json`). |
 | `/tests/` | Pytest suites; keep parity with production code. |
 | `/Makefile` & `setup.ps1` | Developer automation (venv, tests, subtitles, render). |
 | `/llms.txt` | Complementary file containing creative context & tone. |
 | `/subtitles/` | Downloaded `.srt` caption files populated by `fetch_subtitles.py`. |
-| `/scripts/srt_to_markdown.py` | Convert `.srt` captions into Futuroptimist script format (handles italics and emoji) |
-| `/scripts/generate_contrib_heatmap.py` | Build an SVG calendar heatmap of PRs authored in Futuroptimist-owned repos |
-| `/scripts/generate_annual_contributions.py` | Bar chart and CSV of yearly PR counts across Futuroptimist-owned repos |
-| `/scripts/generate_heatmap.py` | Create a 3‑D lines-of-code heatmap with light/dark SVGs |
+| `/src/srt_to_markdown.py` | Convert `.srt` captions into Futuroptimist script format (handles italics and emoji) |
+| `/src/generate_contrib_heatmap.py` | Build an SVG calendar heatmap of PRs authored in Futuroptimist-owned repos |
+| `/src/generate_annual_contributions.py` | Bar chart and CSV of yearly PR counts across Futuroptimist-owned repos |
+| `/src/generate_heatmap.py` | Create a 3‑D lines-of-code heatmap with light/dark SVGs |
 | `/sources/` | Reference files fetched via `collect_sources.py`. |
 | `video_ids.txt` | Canonical list of YouTube IDs referenced by helper scripts. |
-| `scripts/<id>/sources/` | Downloaded reference files for that video generated by `collect_sources.py` (gitignored). |
+| `video_scripts/<id>/sources/` | Downloaded reference files for that video generated by `collect_sources.py` (gitignored). |
 | `INSTRUCTIONS.md` | Extended setup steps and roadmap. |
 
 ## Coding Conventions
@@ -39,7 +40,7 @@ Avoid adding setup or asset instructions there; link to INSTRUCTIONS instead.
 
 ## Script Format
 
-Video scripts (`scripts/YYYYMMDD_slug/script.md`) combine narration and stage
+Video scripts (`video_scripts/YYYYMMDD_slug/script.md`) combine narration and stage
 directions. Use `[NARRATOR]:` for spoken lines and `[VISUAL]:` for b-roll or
 graphics cues. Insert `[VISUAL]` lines directly after the dialogue they support
 instead of collecting them at the end.
@@ -51,12 +52,12 @@ instead of collecting them at the end.
 - Each script folder must include a `metadata.json` file conforming to `schemas/video_metadata.schema.json`. Optional fields like `slug`, `thumbnail`, `transcript_file`, and `summary` enrich automation but aren't required.
 - Each script folder may include a `sources.txt` file with one URL per line. Any downloaded articles or clips are for reference only—check usage rights and cite sources in **APA style** rather than redistributing content.
 - Each script folder may also contain a `footage.md` checklist to track B-roll or CGI shots to gather. Note existing archive vs new footage, and flag generative AI segments so they don't look like "AI slop".
-- Large photos or video files belong in a local `footage/` folder (ignored by git). Run `python scripts/index_local_media.py` whenever assets change to rebuild `footage_index.json` for quick lookup during editing.
-- Run `python scripts/update_transcript_links.py` to sync `transcript_file` paths; with `YOUTUBE_API_KEY` set it also fetches missing captions via YouTube Data API.
+- Large photos or video files belong in a local `footage/` folder (ignored by git). Run `python src/index_local_media.py` whenever assets change to rebuild `footage_index.json` for quick lookup during editing.
+- Run `python src/update_transcript_links.py` to sync `transcript_file` paths; with `YOUTUBE_API_KEY` set it also fetches missing captions via YouTube Data API.
 
 ## Testing & CI
 
-4. Optionally run `make subtitles` and `python scripts/collect_sources.py` to
+4. Optionally run `make subtitles` and `python src/collect_sources.py` to
    download captions and reference files.
 
 The repository includes a simple GitHub Actions workflow (`.github/workflows/02-tests.yml`)
@@ -70,14 +71,14 @@ push or pull request.
 make setup   # venv + deps (or ./setup.ps1)
 make test    # runs `pytest -q`
 # To check coverage locally:
-pytest --cov=./scripts --cov=./tests
+pytest --cov=./src --cov=./tests
 ```
 The Makefile auto-detects Windows vs Unix paths so these commands should work cross-platform.
 If `make setup` fails on your platform, run `python3 -m venv .venv && uv pip install -r requirements.txt` then `pytest -q`.
 If `make test` errors about `.venv/Scripts/python`, use `PATH=.venv/bin:$PATH pytest -q` instead.
 If `yt-dlp` cannot be located during tests, prefix your command with `PATH=.venv/bin:$PATH` so the venv's executables are discoverable.
 
-CI runs `pytest --cov=./scripts --cov=./tests` on every push and pull request targeting `main`. In a future phase it may also run `make subtitles` to verify caption downloads.
+CI runs `pytest --cov=./src --cov=./tests` on every push and pull request targeting `main`. In a future phase it may also run `make subtitles` to verify caption downloads.
 Aim to keep coverage at **100%** so the Codecov badge stays green.
 
 When adding CLI entrypoint tests, stub out any network requests (e.g. patch
@@ -86,7 +87,7 @@ are unreachable.
 
 ## Vision & Workflow
 
-The long-term goal is to make video creation as repeatable as writing code. Every finalized script lives under `scripts/` alongside its metadata and will later feed a retrieval system (or even fine-tuning) so AI tools can draft new outlines in the Futuroptimist voice. Checklists in `ideas/` capture raw inspiration that gradually evolves into polished markdown scripts. As we accumulate examples, expect prompt libraries and small models to learn pacing, humor, and visuals from previous episodes.
+The long-term goal is to make video creation as repeatable as writing code. Every finalized script lives under `video_scripts/` alongside its metadata and will later feed a retrieval system (or even fine-tuning) so AI tools can draft new outlines in the Futuroptimist voice. Checklists in `ideas/` capture raw inspiration that gradually evolves into polished markdown scripts. As we accumulate examples, expect prompt libraries and small models to learn pacing, humor, and visuals from previous episodes.
 
 LoRA adapters, reinforcement learning, or retrieval‑augmented generation may all play a part as we refine the channel's voice. Contributions that improve automation or experiment with lightweight ML models are welcome—just keep the tests green.
 
@@ -98,7 +99,7 @@ When Phase 7 hits (see the roadmap in INSTRUCTIONS.md) an additional `make rende
 
 1. Fork & branch.
 2. `make setup` then `make test`.
-3. `python scripts/scaffold_videos.py` to pull metadata and create dated script folders from `video_ids.txt` (commit new folders).
+3. `python src/scaffold_videos.py` to pull metadata and create dated script folders from `video_ids.txt` (commit new folders).
 4. Optionally run `make subtitles` to verify caption downloads.
 5. Add or update code **and** matching tests.
 6. Commit with descriptive message; open PR.
