+++ src/svg3d.py
