+++ video_scripts/20210917_how-tesla-bots-could-benefit-humanity/script.md
