+++ video_scripts/20211001_why-tesla-s-full-self-driving-is-taking-so-long/metadata.json
