+++ video_scripts/20211102_we-re-probably-in-a-simulation-because-we-make-them-already/script.md
