+++ video_scripts/20250701_indoor-aquariums-tour/metadata.json
